+++ packages/runtime/src/interpreter/opcode.ts
@@ -6,6 +6,7 @@ import {
 	GlobalFields,
 	ITxArrFields,
 	ITxnFields,
+	MathOp,
 	MAX_UINT6,
 	MAX_UINT8,
 	MAX_UINT64,
@@ -17,7 +18,21 @@ import {
 } from "../lib/constants";
 import type { TEALStack } from "../types";
 
-export class Op {
+export abstract class Op {
+	//line number in TEAL file
+	abstract line: number;
+	/**
+	 * executes the opcode and returns its cost if different from 1
+	 * @param stack TEAL stack
+	 * @param op math operator for the ByteOp class
+	 * @returns either the cost or if cost is default (equal to 1) void;
+	 */
+	abstract execute(stack: TEALStack, op?: MathOp): number;
+
+	computeCost(): number {
+		return 1;
+	}
+
 	/**
 	 * assert stack length is atleast minLen
 	 * @param stack TEAL stack