+++ packages/runtime/src/runtime.ts
@@ -226,6 +226,19 @@ export class Runtime {
 	}
 
 	/**
+	 * Ensure no duplicate transaction in group txn
+	 * @param gtxns group transaction
+	 */
+	assertNoDuplicateTransaction(gtxns: EncTx[]): void {
+		for (const txn of gtxns) {
+			const isDuplicate = gtxns.filter((anotherTxn) => anotherTxn.txID === txn.txID).length > 1;
+			if (isDuplicate) {
+				throw new RuntimeError(RUNTIME_ERRORS.TRANSACTION.TRANSACTION_ALREADY_IN_LEDGER);
+			}
+		}
+	}
+
+	/**
 	 * set current round with timestamp for a block
 	 * @param r current round
 	 * @param timestamp block's timestamp
@@ -901,7 +914,8 @@ export class Runtime {
 
 		// validate first and last rounds
 		this.validateTxRound(gtxs);
-
+		// checks if the transactions are not duplicated
+		this.assertNoDuplicateTransaction(gtxs);
 		// initialize context before each execution
 		// Prepare shared space at each execution of transaction/s.
 		// state is a deep copy of store