+++ packages/runtime/test/src/interpreter/opcode.ts
@@ -1,14 +1,16 @@
 import { parsing } from "@algo-builder/web";
+import { assert } from "chai";
 
 import { RUNTIME_ERRORS } from "../../../src/errors/errors-list";
-import { Op } from "../../../src/interpreter/opcode";
+import { Len } from "../../../src/interpreter/opcode-list";
 import { MAX_UINT64, MIN_UINT64 } from "../../../src/lib/constants";
 import { Stack } from "../../../src/lib/stack";
 import type { StackElem } from "../../../src/types";
 import { expectRuntimeError } from "../../helpers/runtime-errors";
 
 describe("Teal Opcodes basic assertions", function () {
-	const op = new Op();
+	//Could be any other opcode. Its only for testing the base abstract class functionalities
+	const op = new Len([], 1);
 
 	it("check uint64 overflow", function () {
 		const max = MAX_UINT64 + BigInt("5");
@@ -43,5 +45,11 @@ describe("Teal Opcodes basic assertions", function () {
 			() => op.assertMinStackLen(stack, stackLen, lineNumber),
 			RUNTIME_ERRORS.TEAL.ASSERT_STACK_LENGTH
 		);
+	});
+
+	it("Should return correct cost", function () {
+		const stack = new Stack<StackElem>();
+		stack.push(parsing.stringToBytes("arg_0"));
+		assert.equal(1, op.execute(stack));
 	});
 });