+++ CHANGELOG.md
@@ -62,6 +62,8 @@ Added:
   console.log("txn1 information: ", receipts[2]);
 ```
 
+- Every opcode class has been updated and now their `execute` method returns its cost. 
+
 
 ### Template improvements
 
@@ -117,7 +119,7 @@ recommended to use the method from `Deployer` rather than the function dirrectly
 - Allow token to be empty.
 - Throw error when issue inner transactions in clear program. Fixed in [#667](https://github.com/scale-it/algo-builder/pull/667).
 - Parameters in `extract*` opcodes can greater than uint8. Fixed in [#666](https://github.com/scale-it/algo-builder/pull/666).
-
+- Restirct duplicate transaction in group transaction. 
 ### Infrastructure
 
 - Updated `setup-master-account` and `sandbox-setup-master-account` commands to run multiple times.